fixing garbsh death time

+++ philosopher.h
@@ -35,16 +35,17 @@ typedef struct philosopher
 	int				n_arg;
 	pthread_mutex_t	mutex;
 }	t_philo;
-int		ft_atoi(const char *str);
-void	sleeping(t_philo	*philos, struct timeval *start1, int c);
-void	ft_usleep(useconds_t time);
-void	create_th(t_philo *philos, int c);
-void	create_p(t_philo *philos, char **arv, int t, int arc);
-void	*h(void *j);
-void	*deathcheck(void *j);
-int		ft_isntdigit(char *arv);
-void	eating(void *j, int c, struct timeval *start, struct timeval *start1);
-int		args_protection(int arc, char **arv);
-void	ft_putnbr_fd(int n);
-int		gettime(struct timeval *start);
+int			ft_atoi(const char *str);
+void		sleeping(t_philo	*philos, struct timeval *start1, int c);
+void		ft_usleep(useconds_t time);
+void		create_th(t_philo *philos, int c);
+void		create_p(t_philo *philos, char **arv, int t, int arc);
+void		*h(void *j);
+void		*deathcheck(void *j);
+int			ft_isntdigit(char *arv);
+void		eating(void *j, int c, struct timeval *start,
+				struct timeval *start1);
+int			args_protection(int arc, char **arv);
+void		ft_putnbr_fd(int n);
+long int	gettime(struct timeval *start);
 #endif
